requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
plotly>=5.18.0
matplotlib>=3.7.0
scipy>=1.11.0
Pillow>=10.0.0
